fix(stella-cli): fleet_cmd's branch watch becomes a sibling module
`fleet_cmd.rs` reached 1538 lines and the 1500-line ceiling is not
grandfathered for it, so `main` went red on `file-size` at `a7a46d3b6`
(#4821) and took every open PR with it.

The seam was already there. `--watch` is the one concern in the file that
begins where every other one has finished: the fan-out has settled, the
dashboard is torn down, the report is written, and what is left is which
branches landed commits, the capped CI wait on each, and the line each
prints. `BranchWatch`, `watch_targets`, `watch_branch` and
`render_watch_line` move to `fleet_cmd/branch_watch.rs` unchanged — same
bodies, same doc comments — beside the `wrapped` and `durability` modules
that were split out on the same reasoning.

`cargo check -p stella-cli` was green while `cargo test` could not build:
`fleet_cmd/tests.rs` calls three of these directly and is `#[cfg(test)]`, so
a check that only compiles the binary cannot see it. That is the same shape
that hid `keymap.rs`'s key-witness `include_str!` during the TUI rename an
hour ago — worth naming twice, because it is the failure mode of verifying a
move with `cargo check`.

fleet_cmd.rs is 1435 lines. `check-file-size` is green; 2288 `stella-cli`
tests pass; `fmt` and `clippy -p stella-cli --all-targets` are clean.

`check-prose` is still red on `agent/goal/goal_wrapped/tests.rs`,
`fleet_cmd/wrapped.rs` and `stella-runtime/README.md` — main's other red,
which #4823 owns. Untouched here rather than fixed twice.

Closes #4821.
